bugfix: handle TypeNotPresentException when processing generic types
Java supports missing classes in type parameters ar run time, due to
the type erasure. However, an attempt to ask for getGenericType() on
such a field would result in TypeNotPresentException.

For example:
class MyBean {
  // fine if compiled with MissingClass which is later removed
  // caused TypeNotPresentException on serialization attempts
  List<MissingClass> list = new ArrayList<>();
}

The problem is fixed by using a backup path (get non-generic type) in
case of exceptions.